consolidating asset icon logic
