+++ extension/src/popup/components/manageAssets/ChooseAsset/styles.scss
@@ -1,0 +1,5 @@
+.ChooseAsset {
+  &__row {
+    display: flex;
+  }
+}
