+++ extension/src/popup/views/ManageAssets/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Switch } from "react-router-dom";
 
 import { AddAsset } from "popup/components/manageAssets/AddAsset";
-import { PreapprovedAssets } from "popup/components/manageAssets/PreapprovedAssets";
+import { ChooseAsset } from "popup/components/manageAssets/ChooseAsset";
 
 import { PrivateKeyRoute } from "popup/Router";
 import { ROUTES } from "popup/constants/routes";
@@ -11,7 +11,7 @@ export const ManageAssets = () => (
   <>
     <Switch>
       <PrivateKeyRoute exact path={ROUTES.manageAssets}>
-        <PreapprovedAssets />
+        <ChooseAsset />
       </PrivateKeyRoute>
       <PrivateKeyRoute exact path={ROUTES.addAsset}>
         <AddAsset />
